Introduce PartOfChain Property To Node Objects
When the final position is reached, the previous position's node will
have its partOfChain value set to true.
For each position part of the queue, test if it has a neighbouring move
that is part of the chain. If so, make it part of the chain, and return
the node + the chain. If not, just return the chain.